HomeTracker - Use ts in the labels to avoid a bug and format only on display
When using the formatted timestamps as labels sometimes the curve will
"jump" e.g if yesterday had records at 21:50 and 21:55 and today has a
record at 21:51 it will be inserted between the records of yesterday
(Actually I'm not sure about that it doesn't really make sense)

Tho this seems to fix the bug with jumping curve.

We use the actual timestamp in all the series and we only format it when
displaying the axis.