Sources - Rework file structure (See full commit message for details)
Motivation: I want to add new apps which don't follow the same svelte
layout as other pages.

Following this comment and the doc I'm creating a group `(apps)` which
will contain the existing apps until now and a second group `(noso)`
which will be a new app with a different layout

https://github.com/sveltejs/kit/issues/9455
https://svelte.dev/docs/kit/advanced-routing#Advanced-layouts-page

According to this SO answer in the future I can use

    git log --follow ./path/to/file

to get the history of the moved files.

https://stackoverflow.com/a/2314757/4194289